tests/wpt/css-tests/css-shapes-1_dev/xhtml1print/shape-outside-invalid-ellipse-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <meta charset="UTF-8" />
    <meta content="dom" name="flags" />
    <meta content="shape-outside declaration is invalid if ellipse shape function has two invalid arguments" name="assert" />
    <title>CSS Shapes Test: Ellipse - invalid 2 arg</title>
    <link href="mailto:sherlock.z0917@gmail.com" rel="author" title="Sherlock Zhong" />
    <link href="mailto:stearns@adobe.com" rel="reviewer" title="Alan Stearns" /> <!-- 11-09-2013 TestTWF Shenzhen -->
    <link href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes" rel="help" />
    <link href="http://www.w3.org/TR/css-shapes-1/#funcdef-ellipse" rel="help" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link href="/resources/testharness.css" rel="stylesheet" />
    <style>
#shape{
    width: 300px;
    height: 300px;
    shape-outside: ellipse(50, 50);
}
    </style>
</head>
<body>

    <div id="log"></div>

    <div id="shape"></div>

    <script>
    var obj = document.getElementById('shape');
    var val = getComputedStyle(obj).shapeOutside;
    test(function() {assert_equals(val, "none", "declaration should not be valid")});
    </script>


</body></html>
